backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
python-dateutil==2.8.2
openai==1.3.7
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
